Extract VertexStream implementations in the app_utils library
* Make the afes library independent from the i/o implementation
details required by the example applications

* CMakeFixes: install the library headers and require dependencies
only when the apps are build
