+++ CMakeLists.txt
@@ -17,7 +17,13 @@ set(CMAKE_EXPORT_COMPILE_COMMANDS ON)
 option(BUILD_APPS "Build applications" ON)
 option(PRECOMPILED_HEADERS "Use precompiled headers for building the applications" OFF)
 
+# Let the module path point to the location where the finders will be generated by
+# the conan cmake_find_package generator
 set(CMAKE_MODULE_PATH ${CMAKE_BINARY_DIR})
+
+# Configure project install folder structure
+# and setup RPATHS for UNIX systems
+include(GNUInstallDirs)
 
 if(APPLE)
   set(basePoint @loader_path)
@@ -25,7 +31,6 @@ else()
   set(basePoint $ORIGIN)
 endif()
 
-include(GNUInstallDirs)
 file(RELATIVE_PATH relDir
   ${CMAKE_CURRENT_BINARY_DIR}/${CMAKE_INSTALL_BINDIR}
   ${CMAKE_CURRENT_BINARY_DIR}/${CMAKE_INSTALL_LIBDIR})
@@ -36,6 +41,5 @@ set(CMAKE_INSTALL_RPATH ${basePoint} ${basePoint}/${relDir})
 include(cmake/find_dependencies.cmake REQUIREED)
 # Setup compiler
 include(cmake/compiler_options.cmake REQUIRED)
-
 
 add_subdirectory(src)
