+++ cmake/find_dependencies.cmake
@@ -1,3 +1,6 @@
+
+if(BUILD_APPS)
 find_package(zlib REQUIRED QUIET)
 find_package(libjpeg REQUIRED QUIET)
 find_package(libpng REQUIRED QUIET)
+endif()
