+++ src/afes/include/afes/VertexStream.h
@@ -29,10 +29,6 @@ DAMAGE.
 #ifndef VERTEX_STREAM_INCLUDED
 #define VERTEX_STREAM_INCLUDED
 
-#include <functional>
-#include "Ply.h"
-#include "Geometry.h"
-
 template< typename Data >
 class InputDataStream
 {
@@ -56,148 +52,5 @@ public:
 	virtual void next( const Data &d ) = 0;
 	virtual void next( const Data *d , size_t count ){ for( size_t i=0 ; i<count ; i++ ) next( d[i] ); }
 };
-
-template< typename Data >
-class TransformedInputDataStream : public InputDataStream< Data >
-{
-	std::function< void ( Data& ) > _xForm;
-	InputDataStream< Data >& _stream;
-public:
-	TransformedInputDataStream( std::function< void ( Data & ) > xForm , InputDataStream< Data > &stream ) : _xForm(xForm) , _stream(stream) {;}
-	virtual void reset( void ){ _stream.reset(); }
-	virtual bool next( Data &d )
-	{
-		bool ret = _stream.next( d );
-		_xForm( d );
-		return ret;
-	}
-};
-
-template< typename Data >
-class TransformedOutputDataStream : public OutputDataStream< Data >
-{
-	std::function< void ( Data & ) > _xForm;
-	OutputDataStream< Data > &_stream;
-public:
-	TransformedOutputDataStream( std::function< void ( Data & ) > xForm , OutputDataStream< Data >& stream ) : _xForm(xForm) , _stream(stream) {;}
-	virtual void reset( void ){ _stream.reset(); }
-	virtual bool next( const Data& p )
-	{
-		Data _p = p;
-		_xForm( _p );
-		return _stream.next( _p );
-	}
-};
-
-template< typename Data >
-class MemoryInputDataStream : public InputDataStream< Data >
-{
-	const Data *_data;
-	size_t _size;
-	size_t _current;
-public:
-	MemoryInputDataStream( size_t size , const Data *data );
-	~MemoryInputDataStream( void );
-	void reset( void );
-	bool next( Data &d );
-};
-
-template< typename Data >
-class MemoryOutputDataStream : public OutputDataStream< Data >
-{
-	Data *_data;
-	size_t _size;
-	size_t _current;
-public:
-	MemoryOutputDataStream( size_t size , Data *data );
-	~MemoryOutputDataStream( void );
-	void next( const Data &d );
-};
-
-template< typename Factory >
-class ASCIIInputDataStream : public InputDataStream< typename Factory::VertexType >
-{
-	typedef typename Factory::VertexType Data;
-	const Factory &_factory;
-	FILE *_fp;
-public:
-	ASCIIInputDataStream( const char* fileName , const Factory &factory );
-	~ASCIIInputDataStream( void );
-	void reset( void );
-	bool next( Data &d );
-};
-
-template< typename Factory >
-class ASCIIOutputDataStream : public OutputDataStream< typename Factory::VertexType >
-{
-	typedef typename Factory::VertexType Data;
-	const Factory &_factory;
-	FILE *_fp;
-public:
-	ASCIIOutputDataStream( const char* fileName , const Factory &factory );
-	~ASCIIOutputDataStream( void );
-	void next( const Data &d );
-};
-
-template< typename Factory >
-class BinaryInputDataStream : public InputDataStream< typename Factory::VertexType >
-{
-	typedef typename Factory::VertexType Data;
-	const Factory &_factory;
-	FILE* _fp;
-public:
-	BinaryInputDataStream( const char* filename , const Factory &factory );
-	~BinaryInputDataStream( void ){ fclose( _fp ) , _fp=NULL; }
-	void reset( void );
-	bool next( Data &d );
-};
-
-template< typename Factory >
-class BinaryOutputDataStream : public OutputDataStream< typename Factory::VertexType >
-{
-	typedef typename Factory::VertexType Data;
-	const Factory &_factory;
-	FILE* _fp;
-public:
-	BinaryOutputDataStream( const char* filename , const Factory &factory );
-	~BinaryOutputDataStream( void ){ fclose( _fp ) , _fp=NULL; }
-	void reset( void ){ fseek( _fp , SEEK_SET , 0 ); }
-	void next( const Data &d );
-};
-
-template< typename Factory >
-class PLYInputDataStream : public InputDataStream< typename Factory::VertexType >
-{
-	typedef typename Factory::VertexType Data;
-	const Factory &_factory;
-	char* _fileName;
-	PlyFile *_ply;
-	std::vector< std::string > _elist;
-	char *_buffer;
-
-	size_t _pCount , _pIdx;
-	void _free( void );
-public:
-	PLYInputDataStream( const char* fileName , const Factory &factory );
-	~PLYInputDataStream( void );
-	void reset( void );
-	bool next( Data &d );
-};
-
-template< typename Factory >
-class PLYOutputDataStream : public OutputDataStream< typename Factory::VertexType >
-{
-	typedef typename Factory::VertexType Data;
-	const Factory &_factory;
-	PlyFile *_ply;
-	size_t _pCount , _pIdx;
-	char *_buffer;
-public:
-	PLYOutputDataStream( const char* fileName , const Factory &factory , size_t count , int fileType=PLY_BINARY_NATIVE );
-	~PLYOutputDataStream( void );
-	void next( const Data &d );
-};
-
-#include "VertexStream.inl"
 
 #endif // VERTEX_STREAM_INCLUDED
