+++ src/apps/AdaptiveTreeVisualization.cpp
@@ -35,10 +35,11 @@ DAMAGE.
 #include <afes/MyMiscellany.h>
 #include <apps_utils/CmdLineParser.h>
 #include <apps_utils/io/Image.h>
+#include <apps_utils/io/Ply.h>
+#include <apps_utils/io/VertexStreamImpl.h>
+#include <apps_utils/io/VertexFactory.h>
 #include <afes/PPolynomial.h>
 #include <afes/FEMTree.h>
-#include <afes/Ply.h>
-#include <afes/VertexFactory.h>
 #include <afes/RegularGrid.h>
 
 cmdLineParameter< char* >
