+++ src/apps/EDTInHeat.cpp
@@ -38,11 +38,11 @@ DAMAGE.
 #include <float.h>
 #include <functional>
 #include <apps_utils/CmdLineParser.h>
+#include <apps_utils/io/Ply.h>
+#include <apps_utils/io/VertexFactory.h>
 #include <afes/MyMiscellany.h>
 #include <afes/PPolynomial.h>
 #include <afes/FEMTree.h>
-#include <afes/Ply.h>
-#include <afes/VertexFactory.h>
 
 MessageWriter messageWriter;
