+++ src/apps/PointInterpolant.cpp
@@ -41,11 +41,12 @@ DAMAGE.
 #include <float.h>
 #include <apps_utils/CmdLineParser.h>
 #include <apps_utils/io/Image.h>
+#include <apps_utils/io/Ply.h>
+#include <apps_utils/io/VertexFactory.h>
+#include <apps_utils/io/VertexStreamImpl.h>
 #include <afes/MyMiscellany.h>
 #include <afes/PPolynomial.h>
 #include <afes/FEMTree.h>
-#include <afes/Ply.h>
-#include <afes/VertexFactory.h>
 #include <afes/RegularGrid.h>
 
 MessageWriter messageWriter;
