+++ src/apps/SurfaceTrimmer.cpp
@@ -35,12 +35,12 @@ DAMAGE.
 #include <float.h>
 #include <algorithm>
 #include <apps_utils/CmdLineParser.h>
+#include <apps_utils/io/Ply.h>
+#include <apps_utils/io/VertexFactory.h>
 #include <afes/FEMTree.h>
 #include <afes/MyMiscellany.h>
 #include <afes/MAT.h>
 #include <afes/Geometry.h>
-#include <afes/Ply.h>
-#include <afes/VertexFactory.h>
 
 MessageWriter messageWriter;
