+++ src/apps_utils/include/apps_utils/io/PlyFile.inl
@@ -36,7 +36,7 @@ WARRANTY OF MERCHANTABILITY OR FITNESS FOR A PARTICULAR PURPOSE.
 #include <math.h>
 #include <string.h>
 #include "PlyFile.h"
-#include "MyMiscellany.h"
+#include <afes/MyMiscellany.h>
 
 const char *type_names[] =
 {
