+++ src/apps_utils/include/apps_utils/io/VertexStreamImpl.inl
